+++ node/lib/debug.js
@@ -102,13 +102,14 @@ debugConnector.prototype.sendCommand = function(obj) {
 
 		var str = JSON.stringify(obj);
 
-		console.log('Send: ');
-		console.log(str);
+		//console.log('Send: ');
+		//console.log(str);
 
         self._waitingForResponse[obj.seq] = obj;
 		self.socket.write( "Content-Length:" + str.length + "\r\n\r\n" + str);
 	} else {
-		console.error('[Node-Debugger] Can\'t send command, not connected!');	
+        //Just ignore it, that is ok
+		//console.error('[Node-Debugger] Can\'t send command, not connected!');
 	}
 };