+++ src/changelogDialog.js
@@ -1,3 +1,4 @@
+/*global define, brackets, $, Mustache */
 define(function (require, exports, module) {
     "use strict";