+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 
 **Please restart Brackets before using Node.js Debugger!**
 [Click here](https://github.com/TheBenji/brackets-node-debugger#how-to-use-it) for more Information
+If you find a bug please open an issue on github!
 
 ## 0.1.3
+* ADD: Save breakpoints on reconnect
+* ADD: Option for auto connect
 * ADD: Settings for Debugger port and host
 * ADD: Changelog
-* ADD: Option for auto connect
+* ADD: Info link to github readme
 * BUGFIX: Breakpoints clear always properly now
 
 ## 0.1.2